pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hyde-my-pii"
version = "0.1.0"
description = "CLI app to detect and replace sensitive data in text, documents, PDFs and images"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "Faker>=24.0.0",
  "pypdf>=4.0.0",
  "python-docx>=1.1.0",
  "pytesseract>=0.3.10",
  "Pillow>=10.0.0",
  "pdf2image>=1.17.0"
]

[project.scripts]
hydemypii = "hydemypii.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
